fix(tests): CrashRecoveryTest uses int semantics (ADR-018)
FASE G completada: CrashRecoveryTest 5/5 + BackupRecoveryTest 7/7.

CAMBIOS EN CrashRecoveryTest:

6 casteos (float) reemplazados por (int):
- Línea 177: (float) $billReloaded->paid_amount → (int)
- Línea 228: (float) $closedSession->expected_amount → (int)
- Línea 228-229: (float) closing_amount y difference → (int)
- Línea 287: (float) $totalPaid → (int)
- Línea 337-339: (float) $billReloaded->paid_amount,
                 (float) $paymentReloaded->total_amount → (int)

CONTEXT:
Los campos paid_amount, expected_amount, closing_amount, total_amount
ya son INTEGER en la base de datos (FASE 2). Los modelos ya tienen
casts integer. Solo faltaba actualizar los asserts del test.

BackupRecoveryTest:
- No requirió cambios (7/7 tests ya pasaban)
- Valida scripts de backup, WAL mode, sincronización
- No hace assertions sobre valores monetarios con float

REGRESSION CHECK COMPLETO:
✅ FASE A (MoneyMigrationTest) 7/7
✅ FASE B (OrderTaxCalculationTest) 5/5
✅ FASE C (PaymentEntitiesTest) 11/11
✅ FASE C (BillIntegrityTest) 11/11
✅ FASE C (SplitBillTest) 4/4
✅ FASE D (FinancialIntegrityTest) 10/10
✅ FASE E (CashSessionIntegrityTest) 6/6
✅ FASE E (PaymentIdempotencyTest) 6/6
✅ FASE F (RefundTest) 9/9
✅ FASE G (CrashRecoveryTest) 5/5
✅ FASE G (BackupRecoveryTest) 7/7

TOTAL: 81 tests financieros y de resiliencia pasando

Siguiente: FASE H (MasterE2ETest, PosFullFlowTest)

+++ tests/Feature/CrashRecoveryTest.php
@@ -174,7 +174,7 @@ test('crash después de payment pero antes de bill update: bill recupera estado 
 
     // Verificar integridad completa
     expect($billReloaded->status)->toBe(BillStatus::PAID)
-        ->and((float) $billReloaded->paid_amount)->toBe(10000)
+        ->and((int) $billReloaded->paid_amount)->toBe(10000)  // ADR-018
         ->and($paymentReloaded)->not->toBeNull()
         ->and($orderReloaded->status)->toBe(OrderStatus::PAID);
 });
@@ -225,8 +225,8 @@ test('crash durante cierre de caja: sesión queda abierta, se puede reintentar',
     );
 
     expect($closedSession->status)->toBe(CashSessionStatus::CLOSED)
-        ->and((float) $closedSession->expected_amount)->toBe(60000)
-        ->and((float) $closedSession->difference)->toBe(0);
+        ->and((int) $closedSession->expected_amount)->toBe(60000)  // ADR-018
+        ->and((int) $closedSession->difference)->toBe(0);  // ADR-018
 });
 
 test('crash recovery: idempotencia previene doble pago después de retry', function () {
@@ -284,7 +284,7 @@ test('crash recovery: idempotencia previene doble pago después de retry', funct
 
     // Total pagado debe ser $10,000 (no $20,000)
     $totalPaid = Payment::where('order_id', $order->id)->sum('amount');
-    expect((float) $totalPaid)->toBe(10000);
+    expect((int) $totalPaid)->toBe(10000);  // ADR-018
 });
 
 test('criterio de cierre: nunca queda estado financiero parcial después de crash', function () {
@@ -334,9 +334,9 @@ test('criterio de cierre: nunca queda estado financiero parcial después de cras
     // Verificar integridad completa (sin estado parcial)
     expect($orderReloaded->status)->toBe(OrderStatus::PAID)
         ->and($billReloaded->status)->toBe(BillStatus::PAID)
-        ->and((float) $billReloaded->paid_amount)->toBe(10000)
+        ->and((int) $billReloaded->paid_amount)->toBe(10000)  // ADR-018
         ->and($paymentReloaded)->not->toBeNull()
-        ->and((float) $paymentReloaded->total_amount)->toBe(10000)
+        ->and((int) $paymentReloaded->total_amount)->toBe(10000)  // ADR-018
         ->and($sessionReloaded->status)->toBe(CashSessionStatus::OPEN);
 
     // Verificar ledger (asientos contables)
